Move AggregationPicker into querying/common/components
