+++ frontend/src/metabase/querying/measures/components/MeasureAggregationPicker/MeasureAggregationPicker.tsx
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from "react";
 import { t } from "ttag";
 
 import ErrorBoundary from "metabase/ErrorBoundary";
-import { AggregationPicker } from "metabase/common/components/AggregationPicker";
+import { AggregationPicker } from "metabase/querying/common/components/AggregationPicker";
 import { ClauseStep } from "metabase/querying/notebook/components/ClauseStep";
 import * as Lib from "metabase-lib";
